fix: package renames
