+++ core/state/factory/state_factory.go
@@ -1,4 +1,4 @@
-package statefactory
+package factory
 
 import (
 	"github.com/NethermindEth/juno/core"
