+++ core/state/factory/state_factory_test.go
@@ -1,4 +1,4 @@
-package statefactory
+package factory
 
 import (
 	"testing"
